feat: typescript implementation
