+++ admin-overrides.yaml
@@ -1,2 +1,2 @@
 interface:
-  customWelcome: Test 123
+  customWelcome: Test 123456
