+++ packages/librechat-admin/package.json
@@ -2,11 +2,13 @@
   "name": "@govgpt/librechat-admin",
   "version": "1.0.0",
   "description": "GovGPT Admin Panel Plugin for LibreChat",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
   "scripts": {
-    "start-dev": "node scripts/start-with-config.js",
-    "prepare-config": "node scripts/pre-start.js",
-    "dev-reload": "node scripts/dev-reload.js"
+    "build": "tsc -p ."
   },
   "dependencies": {
     "js-yaml": "^4.1.0",
@@ -15,5 +17,12 @@
   },
   "peerDependencies": {
     "express": "^4.18.0"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.17",
+    "@types/js-yaml": "^4.0.5",
+    "@types/lodash": "^4.14.195",
+    "@types/node": "^20.0.0",
+    "typescript": "^5.0.4"
   }
 }
