+++ packages/librechat-admin/src/router/index.ts
@@ -1,27 +1,24 @@
-const express = require('express');
-const { getOverrides, updateOverride, applyChanges } = require('../services/configService');
-const path = require('path');
-const fs = require('fs');
+import express, { Router } from 'express';
+import { getOverrides, updateOverride, applyChanges } from '../services/configService';
+import path from 'path';
+import fs from 'fs';
 
 /**
  * Creates an Express router for the GovGPT admin plugin.
- * @param {object} options - Customisation options passed by the consumer.
- * @returns {import('express').Router}
+ * @param options Customisation options passed by the consumer.
  */
-function createRouter(options = {}) {
+function createRouter(options: Record<string, unknown> = {}): Router {
   const router = express.Router();
 
-  // Basic health-check
   router.get('/health', (_req, res) => {
     res.json({ plugin: 'govgpt-admin', status: 'ok' });
   });
 
-  // Replace placeholder endpoints
-  router.get('/config', async (req, res) => {
+  router.get('/config', async (_req, res) => {
     try {
       const overrides = await getOverrides();
       res.json({ overrides });
-    } catch (err) {
+    } catch (err: any) {
       res.status(500).json({ message: err.message });
     }
   });
@@ -32,26 +29,25 @@ function createRouter(options = {}) {
       if (!key) {
         return res.status(400).json({ message: 'key required' });
       }
+      // @ts-ignore – user extension of Express Request
       const overrides = await updateOverride(key, value, req.user?.id);
       res.json({ overrides });
-    } catch (err) {
+    } catch (err: any) {
       res.status(err.status || 500).json({ message: err.message });
     }
   });
 
-  // Apply endpoint - now mainly for restart signaling since config updates are immediate
-  router.post('/config/apply', async (req, res) => {
+  router.post('/config/apply', async (_req, res) => {
     try {
       await applyChanges();
       res.json({ message: 'Restart flag written' });
-    } catch (err) {
+    } catch (err: any) {
       res.status(500).json({ message: err.message });
     }
   });
 
-  // Serve static built admin frontend if present
-  // ../../.. takes us from packages/librechat-admin/router → project root
-  const distPath = path.resolve(__dirname, '..', '..', '..', 'admin-frontend', 'dist');
+  // Locate the admin-frontend/dist folder from the project root, regardless of where this file lives
+  const distPath = path.resolve(process.cwd(), 'admin-frontend', 'dist');
   if (fs.existsSync(distPath)) {
     router.use('/', express.static(distPath));
     router.get('*', (_req, res) => {
@@ -62,4 +58,6 @@ function createRouter(options = {}) {
   return router;
 }
 
-module.exports = createRouter; 
+// @ts-ignore CommonJS default export
+module.exports = createRouter;
+export default createRouter;
